// src/interfaces/approval.ts
export interface Approval {
  id: number;
  status: string;
  application_date: string;
  classification: string;
  application_no: number;
  payday: string;
  applicant: string;
  title: string;
  account_subject_1: string;
  purchase_transportation: string;
  amount: number;
  expense_charge: string;
}